Enable refine widget via prompt.
